Refactor of project structure
