+++ src/schema/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './typedefs';
+export * from './query';
+export * from './mutation';
