+++ src/schema/index.ts
@@ -1,12 +1,13 @@
 import { makeSchema } from 'nexus';
 import { join } from 'path';
-import * as UserSchema from './auth/user';
-import * as VotationSchema from './votation/votation';
-import * as MeetingSchema from './meeting/meetings';
+import * as UserSchema from './auth/';
+import * as VotationSchema from './votation/';
+import * as MeetingSchema from './meeting';
+import * as Enums from './enums';
 import { nexusPrisma } from 'nexus-plugin-prisma';
 
 export const schema = makeSchema({
-    types: [UserSchema, VotationSchema, MeetingSchema], // 1
+    types: [Enums, UserSchema, VotationSchema, MeetingSchema], // 1
     outputs: {
         typegen: join(__dirname, '../__generated__', 'nexus-typegen.ts'), // 2
         schema: join(__dirname, '../__generated__', 'schema.graphql'), // 3
