+++ src/schema/meeting/index.ts
@@ -1,0 +1,2 @@
+export * from './typedefs';
+export * from './query';
