+++ src/schema/votation/index.ts
@@ -1,0 +1,4 @@
+export * from './typedefs';
+export * from './query';
+export * from './mutation';
+export * from './utils';
